Improved warnings suppression.
Accepted Troy's suggestions. In tpl_dump(), cast to int from a write() return value
is performed after checking for INT_MAX.
Also, avoiding use of ssize_t (in favour to size_t or long+cast) to be windows
friendly.

+++ src/tpl.h
@@ -120,12 +120,12 @@ TPL_API tpl_node *tpl_map(char *fmt,...);       /* define tpl using format */
 TPL_API void tpl_free(tpl_node *r);             /* free a tpl map */
 TPL_API int tpl_pack(tpl_node *r, int i);       /* pack the n'th packable */
 TPL_API int tpl_unpack(tpl_node *r, int i);     /* unpack the n'th packable */
-TPL_API ssize_t tpl_dump(tpl_node *r, int mode, ...); /* serialize to mem/file */
+TPL_API int tpl_dump(tpl_node *r, int mode, ...); /* serialize to mem/file */
 TPL_API int tpl_load(tpl_node *r, int mode, ...); /* set mem/file to unpack */
 TPL_API int tpl_Alen(tpl_node *r, int i);      /* array len of packable i */
 TPL_API char* tpl_peek(int mode, ...);         /* sneak peek at format string */
 TPL_API int tpl_gather( int mode, ...);        /* non-blocking image gather */
-TPL_API ssize_t tpl_jot(int mode, ...);            /* quick write a simple tpl */
+TPL_API int tpl_jot(int mode, ...);            /* quick write a simple tpl */
 
 TPL_API tpl_node *tpl_map_va(char *fmt, va_list ap);
